WalletPage-fix-separated stepper into component
